#237 show channel when listing orders
